Separate L4 load balancer code
