+++ pkg/cloudprovider/providers/nifcloud/nifcloud_load_balancer.go
@@ -3,13 +3,9 @@ package nifcloud
 import (
 	"context"
 	"fmt"
-	"sort"
-	"strconv"
 	"strings"
 
 	v1 "k8s.io/api/core/v1"
-	servicehelpers "k8s.io/cloud-provider/service/helpers"
-	"k8s.io/klog/v2"
 )
 
 const (
@@ -60,7 +56,7 @@ const (
 	ServiceAnnotationLoadBalancerHCInterval = "service.beta.kubernetes.io/nifcloud-load-balancer-healthcheck-interval"
 
 	// ServiceAnnotationLoadBalancerType is the annotation that specify using load balancer type
-	// valid values are 'lb' or 'elb'
+	// valid values are 'lb'(default) or 'elb'
 	ServiceAnnotationLoadBalancerType = "service.beta.kubernetes.io/nifcloud-load-balancer-type"
 
 	// ServiceAnnotationLoadBalancerNetworkInterface(1-2) is the annotation that specify network interface of elastic load balancer
@@ -88,19 +84,10 @@ func (c *Cloud) GetLoadBalancer(ctx context.Context, clusterName string, service
 	if isElasticLoadBalancer(service.Annotations) {
 		return c.getElasticLoadBalancer(ctx, clusterName, service)
 	}
-
-	loadBalancerName := c.GetLoadBalancerName(ctx, clusterName, service)
-	loadBalancers, err := c.client.DescribeLoadBalancers(ctx, loadBalancerName)
-	if err != nil {
-		return nil, false, err
+	if isL4LoadBalancer(service.Annotations) {
+		return c.getL4LoadBalancer(ctx, clusterName, service)
 	}
-
-	if len(loadBalancers) == 0 {
-		return nil, false, fmt.Errorf("not found load balancer: %q", loadBalancerName)
-	}
-
-	// service can have many ports, but the load balancer vip is the same
-	return toLoadBalancerStatus(loadBalancers[0].VIP), true, nil
+	return nil, false, fmt.Errorf("the load balancer type is not supported")
 }
 
 // GetLoadBalancerName returns the name of the load balancer
@@ -139,117 +126,15 @@ func (c *Cloud) EnsureLoadBalancer(ctx context.Context, clusterName string, serv
 			return nil, err
 		}
 		return c.ensureElasticLoadBalancer(ctx, loadBalancerName, elb)
-	} else {
-		desire := make([]LoadBalancer, portCount)
-		for i, port := range service.Spec.Ports {
-			// basic load balancer options
-			desire[i].Name = loadBalancerName
-			annotations := service.Annotations
-			if rawBalancingType, ok := annotations[ServiceAnnotationLoadBalancerBalancingType]; ok {
-				balancingType, err := strconv.Atoi(rawBalancingType)
-				if err != nil {
-					return nil, fmt.Errorf(
-						"balancing type %q is invalid for service %q: %v",
-						rawBalancingType, service.GetName(), err,
-					)
-				}
-				desire[i].BalancingType = int32(balancingType)
-			}
-
-			if accountingType, ok := annotations[ServiceAnnotationLoadBalancerAccountingType]; ok {
-				desire[i].AccountingType = accountingType
-			}
-
-			if networkVolume, ok := annotations[ServiceAnnotationLoadBalancerNetworkVolume]; ok {
-				v, err := strconv.Atoi(networkVolume)
-				if err != nil {
-					return nil, fmt.Errorf(
-						"network volume %q is invalid for service %q: %v",
-						networkVolume, service.GetName(), err,
-					)
-				}
-				desire[i].NetworkVolume = int32(v)
-			}
-			if policyType, ok := annotations[ServiceAnnotationLoadBalancerPolicyType]; ok {
-				desire[i].PolicyType = policyType
-			}
-
-			if port.Protocol != v1.ProtocolTCP {
-				return nil, fmt.Errorf("only TCP load balancer is supported")
-			}
-			if port.NodePort == 0 {
-				klog.Errorf("Ignoring port without NodePort defined: %v", port)
-				continue
-			}
-
-			desire[i].LoadBalancerPort = int32(port.Port)
-			desire[i].InstancePort = int32(port.NodePort)
-
-			// health check
-			if strInterval, ok := annotations[ServiceAnnotationLoadBalancerHCInterval]; ok {
-				interval, err := strconv.Atoi(strInterval)
-				if err != nil {
-					return nil, fmt.Errorf(
-						"health check interval %q is invalid for service %q: %v",
-						strInterval, service.GetName(), err,
-					)
-				}
-				desire[i].HealthCheckInterval = int32(interval)
-			} else {
-				desire[i].HealthCheckInterval = defaultHealthCheckInterval
-			}
-
-			if unhealthyThreshold, ok := annotations[ServiceAnnotationLoadBalancerHCUnhealthyThreshold]; ok {
-				t, err := strconv.Atoi(unhealthyThreshold)
-				if err != nil {
-					return nil, fmt.Errorf(
-						"unhealthy threshold %q is invalid for service %q: %v",
-						unhealthyThreshold, service.GetName(), err,
-					)
-				}
-				desire[i].HealthCheckUnhealthyThreshold = int32(t)
-			} else {
-				desire[i].HealthCheckUnhealthyThreshold = defaultHealthCheckUnhealthyThreshold
-			}
-
-			if proto, ok := annotations[ServiceAnnotationLoadBalancerHCProtocol]; ok {
-				switch strings.ToUpper(proto) {
-				case "TCP":
-					desire[i].HealthCheckTarget = fmt.Sprintf("TCP:%d", port.NodePort)
-				case "ICMP":
-					desire[i].HealthCheckTarget = "ICMP"
-				default:
-					return nil, fmt.Errorf(
-						"health check protocol %q is invalid for service %q",
-						proto, service.GetName(),
-					)
-				}
-			} else {
-				desire[i].HealthCheckTarget = fmt.Sprintf("%s:%d", defaultHealthCheckTarget, port.NodePort)
-			}
-
-			// balancing targets
-			desire[i].BalancingTargets = instances
-
-			// filter
-			sourceRanges, err := servicehelpers.GetLoadBalancerSourceRanges(service)
-			if err != nil {
-				return nil, err
-			}
-			filters := []string{}
-			if !servicehelpers.IsAllowAll(sourceRanges) {
-				for cidr := range sourceRanges {
-					if strings.HasSuffix(cidr, "/32") {
-						filters = append(filters, strings.TrimSuffix(cidr, "/32"))
-					} else {
-						filters = append(filters, strings.Replace(cidr, "/32", "", 1))
-					}
-				}
-			}
-			desire[i].Filters = sort.StringSlice(filters)
-		}
-		return c.ensureLoadBalancer(ctx, desire)
 	}
+	if isL4LoadBalancer(service.Annotations) {
+		l4lb, err := NewL4LoadBalancerFromService(loadBalancerName, instances, service)
+		if err != nil {
+			return nil, err
+		}
+		return c.ensureL4LoadBalancer(ctx, loadBalancerName, l4lb)
+	}
+	return nil, fmt.Errorf("the load balancer type is not supported")
 }
 
 // UpdateLoadBalancer updates hosts under the specified load balancer
@@ -257,20 +142,10 @@ func (c *Cloud) UpdateLoadBalancer(ctx context.Context, clusterName string, serv
 	if isElasticLoadBalancer(service.Annotations) {
 		return c.updateElasticLoadBalancer(ctx, clusterName, service, nodes)
 	}
-
-	loadBalancerName := c.GetLoadBalancerName(ctx, clusterName, service)
-
-	loadBalancers, err := c.client.DescribeLoadBalancers(ctx, loadBalancerName)
-	if err != nil {
-		return err
+	if isL4LoadBalancer(service.Annotations) {
+		return c.updateL4LoadBalancer(ctx, clusterName, service, nodes)
 	}
-	if len(loadBalancers) == 0 {
-		return fmt.Errorf("load balancer %q not found", loadBalancerName)
-	}
-
-	_, err = c.EnsureLoadBalancer(ctx, clusterName, service, nodes)
-
-	return err
+	return fmt.Errorf("the load balancer type is not supported")
 }
 
 // EnsureLoadBalancerDeleted deletes the specified load balancer if it exists
@@ -278,215 +153,10 @@ func (c *Cloud) EnsureLoadBalancerDeleted(ctx context.Context, clusterName strin
 	if isElasticLoadBalancer(service.Annotations) {
 		return c.ensureElasticLoadBalancerDeleted(ctx, clusterName, service)
 	}
-
-	loadBalancerName := c.GetLoadBalancerName(ctx, clusterName, service)
-
-	loadBalancers, err := c.client.DescribeLoadBalancers(ctx, loadBalancerName)
-	if err != nil {
-		return err
+	if isL4LoadBalancer(service.Annotations) {
+		return c.ensureL4LoadBalancerDeleted(ctx, clusterName, service)
 	}
-	if len(loadBalancers) == 0 {
-		return fmt.Errorf("load balancer %q already deleted", loadBalancerName)
-	}
-
-	for _, lb := range loadBalancers {
-		klog.Infof("Deleting LoadBalancer %q (%d -> %d)", lb.Name, lb.LoadBalancerPort, lb.InstancePort)
-		if err := c.client.DeleteLoadBalancer(ctx, &lb); err != nil {
-			return fmt.Errorf("failed to delete load balancer: %w", err)
-		}
-	}
-
-	return nil
-}
-
-func (c *Cloud) ensureLoadBalancer(ctx context.Context, desire []LoadBalancer) (*v1.LoadBalancerStatus, error) {
-	if len(desire) == 0 {
-		return nil, fmt.Errorf("desire LoadBalancer length must be larger than 1")
-	}
-
-	loadBalancerName := desire[0].Name
-	current, err := c.client.DescribeLoadBalancers(ctx, loadBalancerName)
-	if err != nil {
-		if strings.Contains(err.Error(), "NotFound") {
-			// create all load balancers
-			var vip string
-			for i, lb := range desire {
-				klog.Infof("Creating LoadBalancer %q (%d -> %d)", lb.Name, lb.LoadBalancerPort, lb.InstancePort)
-				if i == 0 {
-					vip, err = c.client.CreateLoadBalancer(ctx, &lb)
-					if err != nil {
-						return nil, fmt.Errorf("failed to create load balancer: %w", err)
-					}
-				} else {
-					if err := c.client.RegisterPortWithLoadBalancer(ctx, &lb); err != nil {
-						return nil, fmt.Errorf("failed to add port to load balancer: %w", err)
-					}
-				}
-			}
-
-			return toLoadBalancerStatus(vip), nil
-		}
-
-		return nil, fmt.Errorf("failed to describe load balanacer %q: %w", loadBalancerName, err)
-	}
-
-	klog.Infof("desire: %v, current: %v", desire, current)
-
-	loadBalancerResourceChanged := false
-	if len(current) < len(desire) {
-		toCreate := loadBalancerDifferences(desire, current)
-		for _, lb := range toCreate {
-			klog.Infof("Creating LoadBalancer %q (%d -> %d)", lb.Name, lb.LoadBalancerPort, lb.InstancePort)
-			if err := c.client.RegisterPortWithLoadBalancer(ctx, &lb); err != nil {
-				return nil, fmt.Errorf("failed to add port to load balancer: %w", err)
-			}
-			loadBalancerResourceChanged = true
-		}
-	} else if len(current) > len(desire) {
-		toDelete := loadBalancerDifferences(current, desire)
-		for _, lb := range toDelete {
-			klog.Infof("Deleting LoadBalancer %q (%d -> %d)", lb.Name, lb.LoadBalancerPort, lb.InstancePort)
-			if err := c.client.DeleteLoadBalancer(ctx, &lb); err != nil {
-				return nil, fmt.Errorf("failed to delete load balancer: %w", err)
-			}
-			loadBalancerResourceChanged = true
-		}
-	}
-
-	// fetch load balancers again to update latest load balancer info
-	if loadBalancerResourceChanged {
-		current, err = c.client.DescribeLoadBalancers(ctx, loadBalancerName)
-		if err != nil {
-			return nil, fmt.Errorf("failed to describe load balanacer %q: %w", loadBalancerName, err)
-		}
-	}
-
-	klog.Infof("desire: %v, current: %v", desire, current)
-
-	for _, currentLB := range current {
-		desireLB, err := findLoadBalancer(desire, currentLB)
-		if err != nil {
-			return nil, err
-		}
-
-		// reconcile balancing targets
-		toRegister := loadBalancingTargetsDifferences(desireLB.BalancingTargets, currentLB.BalancingTargets)
-		if len(toRegister) > 0 {
-			klog.Infof(
-				"Register instances with load balancer %q (%d -> %d): %v",
-				currentLB.Name, currentLB.LoadBalancerPort, currentLB.InstancePort, toRegister,
-			)
-			if err := c.client.RegisterInstancesWithLoadBalancer(ctx, &currentLB, toRegister); err != nil {
-				return nil, fmt.Errorf("failed to register instances: %w", err)
-			}
-		}
-
-		toDeregister := loadBalancingTargetsDifferences(currentLB.BalancingTargets, desireLB.BalancingTargets)
-		if len(toDeregister) > 0 {
-			klog.Infof(
-				"Deregister instances from load balancer %q (%d -> %d): %v",
-				currentLB.Name, currentLB.LoadBalancerPort, currentLB.InstancePort, toDeregister,
-			)
-			if err := c.client.DeregisterInstancesFromLoadBalancer(ctx, &currentLB, toDeregister); err != nil {
-				return nil, fmt.Errorf("failed to deregister instances: %w", err)
-			}
-		}
-
-		// reconcile filters
-		toAuthorize := filterDifferences(desireLB.Filters, currentLB.Filters)
-		toRevoke := filterDifferences(currentLB.Filters, desireLB.Filters)
-		toSet := []Filter{}
-		for _, addr := range toAuthorize {
-			if addr == filterAnyIPAddresses {
-				continue
-			}
-			toSet = append(toSet, Filter{AddOnFilter: true, IPAddress: addr})
-		}
-		for _, addr := range toRevoke {
-			if addr == filterAnyIPAddresses {
-				continue
-			}
-			toSet = append(toSet, Filter{AddOnFilter: false, IPAddress: addr})
-		}
-		if len(toSet) > 0 {
-			klog.Infof("Applying filter: %v", toSet)
-			if err := c.client.SetFilterForLoadBalancer(ctx, &currentLB, toSet); err != nil {
-				return nil, fmt.Errorf("failed to set filter for load balancer: %w", err)
-			}
-		}
-	}
-
-	return toLoadBalancerStatus(current[0].VIP), nil
-}
-
-func findLoadBalancer(from []LoadBalancer, target LoadBalancer) (*LoadBalancer, error) {
-	for _, lb := range from {
-		if target.Equals(lb) {
-			return &lb, nil
-		}
-	}
-
-	return nil, fmt.Errorf(
-		"target load balancer (%q: %d -> %d) not found",
-		target.Name, target.LoadBalancerPort, target.InstancePort,
-	)
-}
-
-func loadBalancerDifferences(target, other []LoadBalancer) []LoadBalancer {
-	diff := []LoadBalancer{}
-	for _, x := range target {
-		found := false
-		for _, y := range other {
-			if x.Equals(y) {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			diff = append(diff, x)
-		}
-	}
-
-	return diff
-}
-
-func loadBalancingTargetsDifferences(target, other []Instance) []Instance {
-	diff := []Instance{}
-	for _, x := range target {
-		found := false
-		for _, y := range other {
-			if x.Equals(y) {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			diff = append(diff, x)
-		}
-	}
-
-	return diff
-}
-
-func filterDifferences(target, other []string) []string {
-	diff := []string{}
-	for _, x := range target {
-		found := false
-		for _, y := range other {
-			if x == y {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			diff = append(diff, x)
-		}
-	}
-
-	return diff
+	return fmt.Errorf("the load balancer type is not supported")
 }
 
 func toLoadBalancerStatus(vip string) *v1.LoadBalancerStatus {
